requirements.txt
langchain
langchain_community
chainlit
openai
requests
pygithub
python-dotenv
langchain_openai
openai
transformers
torch